+++ src/AlgebraOfGraphics.jl
@@ -35,7 +35,7 @@ export hideinnerdecorations!, deleteemptyaxes!
 export Layer, Layers, ProcessedLayer, ProcessedLayers, zerolayer
 export Entry, AxisEntries
 export renamer, sorter, nonnumeric, verbatim, presorted
-export density, histogram, linear, smooth, expectation, frequency, contours, filled_contours
+export density, histogram, linear, smooth, expectation, frequency, contours, filled_contours, aggregate
 export visual, data, geodata, dims, mapping
 export datetimeticks
 export draw, draw!
@@ -72,6 +72,7 @@ include("transformations/histogram.jl")
 include("transformations/groupreduce.jl")
 include("transformations/frequency.jl")
 include("transformations/expectation.jl")
+include("transformations/aggregate.jl")
 include("transformations/contours.jl")
 include("transformations/filled_contours.jl")
 include("guides/guides.jl")